working on local rules
