+++ local_rules/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "local_rules"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tree-sitter = "0.20.10"
+tree-sitter-lint = { path = ".." }
+tree-sitter-grep = { package = "tree_sitter_lint_tree-sitter-grep", git = "https://github.com/helixbass/tree-sitter-grep", rev = "5fe38b38", version = "0.1.0" }
