+++ local_rules/src/rules/prefer_impl_param.rs
@@ -1,9 +1,7 @@
 use std::sync::Arc;
 
-use proc_macros::rule;
 use tree_sitter::Node;
-
-use crate::{rule::Rule, violation};
+use tree_sitter_lint::{rule, violation, Rule};
 
 #[macro_export]
 macro_rules! assert_node_kind {
@@ -159,10 +157,9 @@ pub fn prefer_impl_param_rule() -> Arc<dyn Rule> {
 
 #[cfg(test)]
 mod tests {
-    use proc_macros::rule_tests;
+    use tree_sitter_lint::{rule_tests, RuleTester};
 
     use super::*;
-    use crate::RuleTester;
 
     #[test]
     fn test_prefer_impl_param_rule() {
