+++ proc_macros/src/builder_args.rs
@@ -1,0 +1,40 @@
+use std::collections::HashMap;
+
+use proc_macro::TokenStream;
+use quote::quote;
+use syn::{
+    parse::{Parse, ParseStream},
+    parse_macro_input, Expr, ExprPath, Ident, Token,
+};
+
+use crate::ArrowSeparatedKeyValuePairs;
+
+struct BuilderArgs {
+    builder_name: ExprPath,
+    args: HashMap<Ident, Expr>,
+}
+
+impl Parse for BuilderArgs {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let builder_name: ExprPath = input.parse()?;
+        input.parse::<Token![,]>()?;
+        let ArrowSeparatedKeyValuePairs {
+            keys_and_values: args,
+        } = input.parse()?;
+        Ok(BuilderArgs { builder_name, args })
+    }
+}
+
+pub fn builder_args(input: TokenStream) -> TokenStream {
+    let BuilderArgs { builder_name, args } = parse_macro_input!(input as BuilderArgs);
+
+    let keys = args.keys();
+    let values = args.values();
+    quote! {
+        #builder_name::default()
+            #(.#keys(#values))*
+            .build()
+            .unwrap()
+    }
+    .into()
+}
